# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Adhams ERP & Logistics Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProduct"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, reserved, dispatched, delivered, cancelled]
        - name: dealer_id
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrder"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrder"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        subcategory:
          type: string
        description:
          type: string
        masterBarcode:
          type: string
        hsnCode:
          type: string
        basePrice:
          type: number
        unit:
          type: string
        dimensions:
          type: string
        weight:
          type: number
        origin:
          type: string
          enum: [manufactured, imported]
        createdAt:
          type: string
      required:
        - id
        - name
        - category
        - masterBarcode
        - hsnCode
        - basePrice
        - unit
        - origin
        - createdAt
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - items
        - total
        - page
        - limit
    CreateProduct:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        subcategory:
          type: string
        description:
          type: string
        hsnCode:
          type: string
        basePrice:
          type: number
        unit:
          type: string
        dimensions:
          type: string
        weight:
          type: number
        origin:
          type: string
          enum: [manufactured, imported]
      required:
        - name
        - category
        - hsnCode
        - basePrice
        - unit
        - origin
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        dealerId:
          type: integer
        dealerName:
          type: string
        status:
          type: string
          enum: [pending, confirmed, reserved, dispatched, delivered, cancelled]
        totalAmount:
          type: number
        advancePaid:
          type: number
        balanceAmount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderNumber
        - dealerId
        - dealerName
        - status
        - totalAmount
        - advancePaid
        - balanceAmount
        - items
        - createdAt
        - updatedAt
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
      required:
        - productId
        - productName
        - quantity
        - unitPrice
        - totalPrice
    OrderListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - items
        - total
        - page
        - limit
    CreateOrder:
      type: object
      properties:
        dealerId:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
              unitPrice:
                type: number
            required:
              - productId
              - quantity
              - unitPrice
        advancePaid:
          type: number
      required:
        - dealerId
        - items
    UpdateOrder:
      type: object
      properties:
        status:
          type: string
        advancePaid:
          type: number
...